requirements.txt
aiofiles
aiosqlite
fastapi
openpyxl
python-dotenv
python-multipart
uvicorn
